Re-plot v0.37.x baseline graphs with new script (#395)
Contributes to #93 and #41

[Rendered](https://github.com/cometbft/cometbft/blob/sergio/qa_0.37.x-alpha3_baseline_plots/docs/qa/v037/CometBFT.md)

Complementary to #376 (and analogous to what we did for CometBFT `v0.34.x` QA) this PR re-plots the baseline results, obtained in Oct 2022, using the newly developed python scripts.
The goal is to ease visual comparison between CometBFT `v0.37.x` and Tendermint Core baseline results when looking at the report.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use [unclog](https://github.com/informalsystems/unclog) to manage our changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code comments